badge,cursoreffect,loader,emojireactions,sharebuttons 컴포넌트화
